+++ hphp/runtime/ext/async_mysql/ext_async_mysql.cpp
@@ -311,7 +311,7 @@ HHVM_METHOD(AsyncMysqlConnectionOptions, setConnectTcpTimeout, int64_t timeout) 
 }
 
 static void
-HHVM_METHOD(AsyncMysqlConnectionOptions, setConnectAttempts, int32_t attempts) {
+HHVM_METHOD(AsyncMysqlConnectionOptions, setConnectAttempts, int64_t attempts) {
   auto* data = Native::data<AsyncMysqlConnectionOptions>(this_);
   data->m_conn_opts.setConnectAttempts(attempts);
 }
@@ -524,7 +524,7 @@ Object HHVM_STATIC_METHOD(
     AsyncMysqlClient,
     connect,
     const String& host,
-    int port,
+    int64_t port,
     const String& dbname,
     const String& user,
     const String& password,
@@ -567,7 +567,7 @@ Object HHVM_STATIC_METHOD(
     AsyncMysqlClient,
     connectWithOpts,
     const String& host,
-    int port,
+    int64_t port,
     const String& dbname,
     const String& user,
     const String& password,
@@ -592,7 +592,7 @@ Object HHVM_STATIC_METHOD(
     connectAndQuery,
     const Variant& queries,
     const String& host,
-    int port,
+    int64_t port,
     const String& dbname,
     const String& user,
     const String& password,
@@ -741,7 +741,7 @@ static Object HHVM_METHOD(
     AsyncMysqlConnectionPool,
     connect,
     const String& host,
-    int port,
+    int64_t port,
     const String& dbname,
     const String& user,
     const String& password,
@@ -786,7 +786,7 @@ static Object HHVM_METHOD(
     AsyncMysqlConnectionPool,
     connectWithOpts,
     const String& host,
-    int port,
+    int64_t port,
     const String& dbname,
     const String& user,
     const String& password,
@@ -814,7 +814,7 @@ static Object HHVM_METHOD(
     connectAndQuery,
     const Variant& queries,
     const String& host,
-    int port,
+    int64_t port,
     const String& dbname,
     const String& user,
     const String& password,
@@ -1084,7 +1084,7 @@ static bool HHVM_METHOD(AsyncMysqlConnection, isSSL) {
   return data->m_conn->isSSL();
 }
 
-static int HHVM_METHOD(AsyncMysqlConnection, warningCount) {
+static int64_t HHVM_METHOD(AsyncMysqlConnection, warningCount) {
   auto* data = Native::data<AsyncMysqlConnection>(this_);
 
   int count = 0;
@@ -1101,7 +1101,7 @@ static String HHVM_METHOD(AsyncMysqlConnection, host) {
   return data->m_host;
 }
 
-static int HHVM_METHOD(AsyncMysqlConnection, port) {
+static int64_t HHVM_METHOD(AsyncMysqlConnection, port) {
   auto* data = Native::data<AsyncMysqlConnection>(this_);
   return data->m_port;
 }
@@ -1260,7 +1260,7 @@ Object AsyncMysqlErrorResult::newInstance(std::shared_ptr<am::Operation> op,
 
 EXTENDS_ASYNC_MYSQL_RESULT(AsyncMysqlErrorResult)
 
-static int HHVM_METHOD(AsyncMysqlErrorResult, mysql_errno) {
+static int64_t HHVM_METHOD(AsyncMysqlErrorResult, mysql_errno) {
   auto* data = Native::data<AsyncMysqlErrorResult>(this_);
   return data->m_op->mysql_errno();
 }
@@ -1316,7 +1316,7 @@ void AsyncMysqlQueryErrorResult::create(std::shared_ptr<am::Operation> op,
   m_query_results = results;
 }
 
-static int HHVM_METHOD(AsyncMysqlQueryErrorResult, numSuccessfulQueries) {
+static int64_t HHVM_METHOD(AsyncMysqlQueryErrorResult, numSuccessfulQueries) {
   auto* data = Native::data<AsyncMysqlQueryErrorResult>(this_);
   return std::dynamic_pointer_cast<am::FetchOperation>(data->m_parent.m_op)
       ->numQueriesExecuted();
