+++ hphp/runtime/ext/process/ext_process.h
@@ -24,7 +24,7 @@ namespace HPHP {
 ///////////////////////////////////////////////////////////////////////////////
 
 int64_t HHVM_FUNCTION(pcntl_alarm,
-                      int seconds);
+                      int64_t seconds);
 void HHVM_FUNCTION(pcntl_exec,
                    const String& path,
                    const Array& args = null_array,
@@ -32,31 +32,31 @@ void HHVM_FUNCTION(pcntl_exec,
 
 int64_t HHVM_FUNCTION(pcntl_fork);
 Variant HHVM_FUNCTION(pcntl_getpriority,
-                      int pid = 0,
-                      int process_identifier = 0);
+                      int64_t pid = 0,
+                      int64_t process_identifier = 0);
 bool HHVM_FUNCTION(pcntl_setpriority,
-                   int priority,
-                   int pid = 0,
-                   int process_identifier = 0);
+                   int64_t priority,
+                   int64_t pid = 0,
+                   int64_t process_identifier = 0);
 
 bool HHVM_FUNCTION(pcntl_signal,
-                   int signo,
+                   int64_t signo,
                    const Variant& handler,
                    bool restart_syscalls = true);
 bool HHVM_FUNCTION(pcntl_sigprocmask,
-                   int how,
+                   int64_t how,
                    const Array& set,
                    Array& oldset);
 int64_t HHVM_FUNCTION(pcntl_wait,
                       int64_t& status,
-                      int options = 0);
+                      int64_t options = 0);
 int64_t HHVM_FUNCTION(pcntl_waitpid,
-                      int pid,
+                      int64_t pid,
                       int64_t& status,
-                      int options = 0);
+                      int64_t options = 0);
 
 int64_t HHVM_FUNCTION(pcntl_wexitstatus,
-                      int status);
+                      int64_t status);
 
 /**
  * Process pending signals flagged earlier.
@@ -65,16 +65,15 @@ bool HHVM_FUNCTION(pcntl_signal_dispatch);
 
 // status querying
 bool HHVM_FUNCTION(pcntl_wifexited,
-                   int status);
+                   int64_t status);
 bool HHVM_FUNCTION(pcntl_wifsignaled,
-                   int status);
+                   int64_t status);
 bool HHVM_FUNCTION(pcntl_wifstopped,
-                   int status);
+                   int64_t status);
 int64_t HHVM_FUNCTION(pcntl_wstopsig,
-                      int status);
+                      int64_t status);
 int64_t HHVM_FUNCTION(pcntl_wtermsig,
-                      int status);
+                      int64_t status);
 
 ///////////////////////////////////////////////////////////////////////////////
 }
-
